+++ aeon/forecasting/__init__.py
@@ -4,10 +4,8 @@ __all__ = [
     "BaseForecaster",
     "NaiveForecaster",
     "RegressionForecaster",
-    "TVPForecaster",
 ]
 
 from aeon.forecasting._naive import NaiveForecaster
 from aeon.forecasting._regression import RegressionForecaster
-from aeon.forecasting._tvp import TVPForecaster
 from aeon.forecasting.base import BaseForecaster
