--- aeon/forecasting/_loss_functions.py
@@ -1,40 +1,0 @@
-"""Loss functions for optimiser."""
-
-import numpy as np
-from numba import njit
-
-from aeon.forecasting._extract_paras import _extract_arma_params
-
-LOG_2PI = 1.8378770664093453
-
-
-@njit(cache=True, fastmath=True)
-def _arima_fit(params, data, model):
-    """Calculate the AIC of an ARIMA model given the parameters."""
-    formatted_params = _extract_arma_params(params, model)  # Extract parameters
-
-    # Initialize residuals
-    n = len(data)
-    residuals = np.zeros(n)
-    c = formatted_params[0][0] if model[0] else 0
-    p = model[1]
-    q = model[2]
-    for t in range(n):
-        ar_term = 0.0
-        max_ar = min(p, t)
-        for j in range(max_ar):
-            ar_term += formatted_params[1, j] * data[t - j - 1]
-        ma_term = 0.0
-        max_ma = min(q, t)
-        for j in range(max_ma):
-            ma_term += formatted_params[2, j] * residuals[t - j - 1]
-        y_hat = c + ar_term + ma_term
-        residuals[t] = data[t] - y_hat
-    sse = 0.0
-    start = max(p, q)
-    for i in range(start, n):
-        sse += residuals[i] * residuals[i]
-    variance = sse / (n - start)
-    likelihood = (n - start) * (LOG_2PI + np.log(variance) + 1.0)
-    k = len(params)
-    return likelihood + 2 * k
