+++ aeon/forecasting/stats/__init__.py
@@ -3,7 +3,9 @@
 __all__ = [
     "ETS",
     "ARIMA",
+    "TVPForecaster",
 ]
 
 from aeon.forecasting.stats._arima import ARIMA
 from aeon.forecasting.stats._ets import ETS
+from aeon.forecasting.stats._tvp import TVPForecaster
